requirements.txt
Django
django-debug-toolbar
django-elasticsearch-dsl
django-mptt
-e git+https://github.com/uktrade/directory-signature-auth.git@0.1.0#egg=directory-signature-auth
-e git+https://github.com/uktrade/directory-header-footer.git@v0.4.0#egg=directory_header_footer
-e git+https://github.com/uktrade/directory-constants.git@v3.1.0#egg=directory_constants
-e git+https://github.com/uktrade/directory-api-client.git@v7.1.0#egg=directory_api_client
-e git+https://github.com/uktrade/directory-sso-api-client.git@0.0.1#egg=directory-sso-api-client
dj-database-url
boto3
django-storages
whitenoise
gunicorn
